email_handler/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pymongo